move editor classes into editor module
