+++ app/src/main/java/com/github/kitakkun/noteapp/navigation/PageRouter.kt
@@ -5,8 +5,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
+import com.github.kitakkun.noteapp.editor.editor.EditorPage
 import com.github.kitakkun.noteapp.finder.FinderPage
-import com.github.kitakkun.noteapp.ui.page.editor.EditorPage
 import org.koin.androidx.compose.koinViewModel
 import org.koin.core.parameter.parametersOf
