--- app/src/main/java/com/github/kitakkun/noteapp/ui/page/editor/dialog/colorpicker/ColorPickerDialog.kt
@@ -1,133 +1,0 @@
-package com.github.kitakkun.noteapp.ui.page.editor.dialog.colorpicker
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.Orientation
-import androidx.compose.foundation.gestures.scrollable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.Button
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
-import com.github.kitakkun.noteapp.R
-import com.github.kitakkun.noteapp.customview.preview.PreviewContainer
-import com.github.skydoves.colorpicker.compose.AlphaSlider
-import com.github.skydoves.colorpicker.compose.AlphaTile
-import com.github.skydoves.colorpicker.compose.BrightnessSlider
-import com.github.skydoves.colorpicker.compose.HsvColorPicker
-import com.github.skydoves.colorpicker.compose.rememberColorPickerController
-
-@Composable
-fun ColorPickerDialog(
-    onDismiss: () -> Unit,
-    onCancel: () -> Unit,
-    onColorConfirm: (Color) -> Unit,
-) {
-    val controller = rememberColorPickerController()
-    var hexCode by remember { mutableStateOf("") }
-    var textColor by remember { mutableStateOf(Color.Transparent) }
-    Dialog(onDismissRequest = onDismiss) {
-        Column(
-            modifier = Modifier
-                .background(MaterialTheme.colorScheme.surface)
-                .scrollable(
-                    state = rememberScrollState(),
-                    orientation = Orientation.Vertical
-                )
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.SpaceBetween,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(4.dp),
-            ) {
-                IconButton(onClick = onCancel) {
-                    Icon(imageVector = Icons.Default.Close, contentDescription = null)
-                }
-                Button(onClick = { onColorConfirm(controller.selectedColor.value) }) {
-                    Text(text = stringResource(R.string.ok))
-                    Spacer(modifier = Modifier.width(4.dp))
-                    Icon(imageVector = Icons.Default.Check, contentDescription = null)
-                }
-            }
-            Text(
-                text = "#$hexCode",
-                color = textColor,
-                fontSize = 16.sp,
-                fontWeight = FontWeight.Bold,
-                modifier = Modifier.align(Alignment.CenterHorizontally)
-            )
-
-            AlphaTile(
-                modifier = Modifier
-                    .size(80.dp)
-                    .clip(RoundedCornerShape(6.dp))
-                    .align(Alignment.CenterHorizontally),
-                controller = controller
-            )
-            HsvColorPicker(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(450.dp)
-                    .padding(10.dp),
-                controller = controller,
-                onColorChanged = { colorEnvelope ->
-                    hexCode = colorEnvelope.hexCode
-                    textColor = colorEnvelope.color
-                }
-            )
-            AlphaSlider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(10.dp)
-                    .height(35.dp),
-                controller = controller,
-            )
-            BrightnessSlider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(10.dp)
-                    .height(35.dp),
-                controller = controller,
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun ColorPickerDialogPreview() = PreviewContainer {
-    ColorPickerDialog(
-        onDismiss = {},
-        onCancel = {},
-        onColorConfirm = {},
-    )
-}
