--- app/src/main/java/com/github/kitakkun/noteapp/ui/page/editor/ext/TextRange.kt
@@ -1,8 +1,0 @@
-package com.github.kitakkun.noteapp.ui.page.editor.ext
-
-import androidx.compose.ui.text.TextRange
-
-fun TextRange.toValidOrder(): TextRange = when {
-    reversed -> TextRange(end, start)
-    else -> this
-}
