+++ editor/src/main/java/com/github/kitakkun/noteapp/editor/editor/EditorModule.kt
@@ -1,0 +1,18 @@
+package com.github.kitakkun.noteapp.editor.editor
+
+import androidx.navigation.NavController
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.core.module.dsl.factoryOf
+import org.koin.dsl.module
+
+val editorModule = module {
+    viewModel { (documentId: String?, navController: NavController) ->
+        EditorViewModel(
+            documentId = documentId,
+            documentRepository = get(),
+            navController = navController,
+            historyManager = get(),
+        )
+    }
+    factoryOf(::EditHistoryManager)
+}
