+++ editor/src/main/java/com/github/kitakkun/noteapp/editor/editor/editmodel/EditHistory.kt
@@ -1,0 +1,11 @@
+package com.github.kitakkun.noteapp.editor.editor.editmodel
+
+import androidx.compose.ui.text.input.TextFieldValue
+import com.github.kitakkun.noteapp.data.model.BaseStyleAnchor
+import com.github.kitakkun.noteapp.data.model.OverrideStyleAnchor
+
+data class EditHistory(
+    val content: TextFieldValue,
+    val baseStyleAnchors: List<BaseStyleAnchor>,
+    val overrideStyleAnchors: List<OverrideStyleAnchor>,
+)
