+++ editor/src/main/java/com/github/kitakkun/noteapp/editor/editor/editmodel/EditorConfig.kt
@@ -1,0 +1,11 @@
+package com.github.kitakkun.noteapp.editor.editor.editmodel
+
+import androidx.compose.ui.graphics.Color
+import com.github.kitakkun.noteapp.data.model.BaseStyle
+
+data class EditorConfig(
+    val baseStyle: BaseStyle = BaseStyle.Body,
+    val isBold: Boolean = false,
+    val isItalic: Boolean = false,
+    val color: Color = Color.Unspecified,
+)
