+++ editor/src/main/java/com/github/kitakkun/noteapp/editor/editor/ext/TextRange.kt
@@ -1,0 +1,8 @@
+package com.github.kitakkun.noteapp.editor.editor.ext
+
+import androidx.compose.ui.text.TextRange
+
+fun TextRange.toValidOrder(): TextRange = when {
+    reversed -> TextRange(end, start)
+    else -> this
+}
